# coffee-be/static/swagger.json
{
	"components": {
...
		"schemas": {
...
			"FieldDefinition": {
				"description": "Definition of an entry field",
				"properties": {
					"label": {
						"type": "string",
						"description": "Label of an entry field"
					},
					"type": {
						"anyOf": [
							{
								"type": "string",
								"enum": [
									"string"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"text"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"number"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"integer"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"date"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"object"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"array"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"price"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"exchange_rate"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"timestamp"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"file"
								],
								"nullable": false
							}
						],
						"description": "Field type (data taype)"
					},
					"required": {
						"type": "boolean",
						"description": "Field is required or not"
					},
					"mandatory": {
						"type": "boolean",
						"description": "Field is mandatory for creation."
					},
					"requiredOnQuote": {
						"type": "boolean",
						"description": "Field is required if related to quote order."
					},
					"stringValue": {
						"type": "string",
						"description": "String value of the field, if the type is 'text', 'date' or 'timestamp'"
					},
					"numericValue": {
						"type": "number",
						"format": "double",
						"description": "Numeric value if the type is 'number'"
					},
					"objectValue": {
						"additionalProperties": true,
						"description": "Object value for type 'object'"
					},
					"fileMultiplicity": {
						"anyOf": [
							{
								"type": "number",
								"format": "double"
							},
							{
								"type": "string",
								"enum": [
									"any"
								],
								"nullable": false
							}
						],
						"description": "Number of files if the type is 'file'. For unlimited number use 'any'."
					},
					"files": {
						"items": {
							"$ref": "#/components/schemas/ChainFileInfo"
						},
						"type": "array",
						"description": "List of files (ChainFileInfo)"
					}
				},
				"required": [
					"label",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ChainCompanyProductRole": {
				"properties": {
					"companyId": {
						"type": "number",
						"format": "double"
					},
					"role": {
						"type": "string"
					}
				},
				"required": [
					"companyId",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ChainProduct": {
				"description": "Represents general product definition of a product. Example: Coffee Angelique (without specific packaging, quantity, ...)\nIt DOES NOT represent an example of a product unit one can buy (e.g. a type of product unit of 50g coffee bag of brand Angelique)\nIt DOES NOT represent any particular product instance/shippment/bag ... (e.g. A particular 50g bag of brand Anguelique in some facility\nsuch that one can actually get it.)\nIt only represents product description with core properties.\nSellable (orderable) types of a product of this kind are described by the class ChainSemiProduct.\nInstances of semi products are described by ChainStockOrder.",
				"properties": {
					"docType": {
						"type": "string",
						"nullable": true
					},
					"_id": {
						"type": "string",
						"nullable": true
					},
					"_rev": {
						"type": "string",
						"nullable": true
					},
					"created": {
						"type": "string",
						"nullable": true,
						"description": "Timestamp of creation"
					},
					"lastChange": {
						"type": "string",
						"nullable": true,
						"description": "Timestamp of last change"
					},
					"userCreatedId": {
						"type": "string",
						"nullable": true,
						"description": "Id of user that created the document."
					},
					"userChangedId": {
						"type": "string",
						"nullable": true,
						"description": "Id of user that changed the document."
					},
					"id": {
						"type": "number",
						"format": "double",
						"description": "ProductId from afApp database"
					},
					"organizationRoles": {
						"items": {
							"$ref": "#/components/schemas/ChainCompanyProductRole"
						},
						"type": "array",
						"nullable": true,
						"description": "List of roles of companies on the product value chain"
					},
					"companyId": {
						"type": "number",
						"format": "double",
						"nullable": true,
						"description": "af coffie company id"
					},
					"organizationId": {
						"type": "string",
						"description": "Chain organization id matching companyId, set automatically from companyId on read or write."
					},
					"organization": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ChainOrganization"
							},
							{
								"type": "number",
								"enum": [
									null
								],
								"nullable": true
							}
						],
						"description": "Chain organization matching companyId, set automatically."
					},
					"description": {
						"type": "string",
						"nullable": true,
						"description": "product description"
					},
					"howToUse": {
						"type": "string",
						"nullable": true,
						"description": "how to Use / Recipes - Describe the best way to use the product (e.g. recipes, how to apply the product...)"
					},
					"ingredients": {
						"type": "string",
						"nullable": true,
						"description": "ingredients - list the ingredients in the product and describe their properties"
					},
					"keyMarketsShare": {
						"properties": {},
						"additionalProperties": {
							"type": "number",
							"format": "double"
						},
						"type": "object",
						"nullable": true,
						"description": "Key Markets, market name - share number map"
					},
					"labels": {
						"items": {
							"$ref": "#/components/schemas/ApiProductLabelValues"
						},
						"type": "array",
						"description": "labels"
					},
					"name": {
						"type": "string",
						"nullable": true,
						"description": "product name"
					},
					"nutritionalValue": {
						"type": "string",
						"nullable": true,
						"description": "nutritional Value - list the nutritional value of the product"
					},
					"origin": {
						"$ref": "#/components/schemas/ApiProductOrigin"
					},
					"photo": {
						"$ref": "#/components/schemas/ApiDocument"
					},
					"process": {
						"$ref": "#/components/schemas/ApiProcess"
					},
					"responsibility": {
						"$ref": "#/components/schemas/ApiResponsibility"
					},
					"sustainability": {
						"$ref": "#/components/schemas/ApiSustainability"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ChainProcessingAction": {
				"properties": {
					"docType": {
						"type": "string",
						"nullable": true
					},
					"_id": {
						"type": "string",
						"nullable": true
					},
					"_rev": {
						"type": "string",
						"nullable": true
					},
					"created": {
						"type": "string",
						"nullable": true,
						"description": "Timestamp of creation"
					},
					"lastChange": {
						"type": "string",
						"nullable": true,
						"description": "Timestamp of last change"
					},
					"userCreatedId": {
						"type": "string",
						"nullable": true,
						"description": "Id of user that created the document."
					},
					"userChangedId": {
						"type": "string",
						"nullable": true,
						"description": "Id of user that changed the document."
					},
					"productId": {
						"type": "string",
						"description": "id reference of the relevant ChainProduct"
					},
					"organizationId": {
						"type": "string",
						"description": "id reference of the relevant ChainOrganization"
					},
					"name": {
						"type": "string",
						"description": "Name of the processing action"
					},
					"description": {
						"type": "string",
						"description": "Description of the processing action"
					},
					"inputSemiProductId": {
						"type": "string",
						"description": "Input semi-product id"
					},
					"outputSemiProductId": {
						"type": "string",
						"description": "Output semi-product id"
					},
					"requiredFields": {
						"items": {
							"$ref": "#/components/schemas/FieldDefinition"
						},
						"type": "array",
						"description": "Required fields"
					},
					"inputSemiProduct": {
						"$ref": "#/components/schemas/ChainSemiProduct",
						"description": "Input semi-product on get only"
					},
					"outputSemiProduct": {
						"$ref": "#/components/schemas/ChainSemiProduct",
						"description": "Output semi-product on get only"
					},
					"requiredDocTypeIds": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "List of required document types for processing ids"
					},
					"requiredDocTypeIdsWithRequired": {
						"items": {
							"$ref": "#/components/schemas/DocTypeIdsWithRequired"
						},
						"type": "array",
						"description": "List of required document types for processing ids"
					},
					"requiredDocTypes": {
						"items": {
							"$ref": "#/components/schemas/ChainProcessingEvidenceType"
						},
						"type": "array",
						"description": "List of required document types for processing ids"
					},
					"repackedOutputs": {
						"type": "boolean",
						"nullable": true,
						"description": "If transaction output has multiple outputSemiProducts"
					},
					"maxOutputWeight": {
						"type": "number",
						"format": "double",
						"nullable": true,
						"description": "Max weight of one of the multiple outputSemiProducts (required if repackedOutputs)"
					},
					"type": {
						"anyOf": [
							{
								"type": "string",
								"enum": [
									"PROCESSING"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"SHIPMENT"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"TRANSFER"
								],
								"nullable": false
							}
						],
						"description": "Type of processing transaction.\nPROCESSING: many-to-many semi products, consumed and produced quantities not connected\nSHIPMENT: same semiproduct. Acts as an order of the same quantity from target facility"
					},
					"prefix": {
						"type": "string",
						"description": "Prefix. Used to build internal lot number names."
					},
					"publicTimelineLabel": {
						"type": "string",
						"nullable": true,
						"description": "Public timeline name. If not null, the processing order with this action is shown on public timeline."
					},
					"publicTimelineLocation": {
						"type": "string",
						"nullable": true,
						"description": "Public timeline location."
					},
					"publicTimelineIcon": {
						"anyOf": [
							{
								"type": "string",
								"enum": [
									"SHIP"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"LEAF"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"WAREHOUSE"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"QRCODE"
								],
								"nullable": false
							},
							{
								"type": "string",
								"enum": [
									"OTHER"
								],
								"nullable": false
							},
							{
								"type": "number",
								"enum": [
									null
								],
								"nullable": true
							}
						],
						"description": "Icon type in public timeline"
					}
				},
				"required": [
					"productId",
					"organizationId",
					"name",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiDefaultResponseStatusEnum": {
				"anyOf": [
					{
						"type": "string",
						"enum": [
							"OK"
						],
						"nullable": false
					},
					{
						"type": "string",
						"enum": [
							"ERROR"
						],
						"nullable": false
					},
					{
						"type": "string",
						"enum": [
							"REQUEST_BODY_ERROR"
						],
						"nullable": false
					},
					{
						"type": "string",
						"enum": [
							"VALIDATION_ERROR"
						],
						"nullable": false
					},
					{
						"type": "string",
						"enum": [
							"TOO_MANY_REQUESTS"
						],
						"nullable": false
					},
					{
						"type": "string",
						"enum": [
							"UNAUTHORIZED"
						],
						"nullable": false
					},
					{
						"type": "string",
						"enum": [
							"AUTH_ERROR"
						],
						"nullable": false
					},
					{
						"type": "string",
						"enum": [
							"UPSTREAM_HTTP_ERROR"
						],
						"nullable": false
					},
					{
						"type": "string",
						"enum": [
							"INVALID_REQUEST"
						],
						"nullable": false
					},
					{
						"type": "string",
						"enum": [
							"NOT_IMPLEMENTED"
						],
						"nullable": false
					}
				],
				"description": "All possible values of status."
			},
			"ApiValidationErrorDetails": {
				"properties": {
					"className": {
						"type": "string"
					},
					"fieldErrors": {
						"properties": {},
						"additionalProperties": {
							"type": "string"
						},
						"type": "object"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_any_": {
				"properties": {
					"data": {
						"additionalProperties": true
					},
					"errorDetails": {
						"type": "string",
						"description": "Optional details for unexpected error responses."
					},
					"errorMessage": {
						"type": "string",
						"description": "Simple message to explain client developers the reason for error."
					},
					"status": {
						"$ref": "#/components/schemas/ApiDefaultResponseStatusEnum",
						"description": "Response status. OK for successful reponses."
					},
					"validationErrorDetails": {
						"$ref": "#/components/schemas/ApiValidationErrorDetails"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_ChainProcessingAction_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/ChainProcessingAction"
					},
					"errorDetails": {
						"type": "string",
						"description": "Optional details for unexpected error responses."
					},
					"errorMessage": {
						"type": "string",
						"description": "Simple message to explain client developers the reason for error."
					},
					"status": {
						"$ref": "#/components/schemas/ApiDefaultResponseStatusEnum",
						"description": "Response status. OK for successful reponses."
					},
					"validationErrorDetails": {
						"$ref": "#/components/schemas/ApiValidationErrorDetails"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_PaginatedList_ChainProcessingAction__": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/PaginatedList_ChainProcessingAction_"
					},
					"errorDetails": {
						"type": "string",
						"description": "Optional details for unexpected error responses."
					},
					"errorMessage": {
						"type": "string",
						"description": "Simple message to explain client developers the reason for error."
					},
					"status": {
						"$ref": "#/components/schemas/ApiDefaultResponseStatusEnum",
						"description": "Response status. OK for successful reponses."
					},
					"validationErrorDetails": {
						"$ref": "#/components/schemas/ApiValidationErrorDetails"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_PaginatedList_ChainProduct__": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/PaginatedList_ChainProduct_"
					},
					"errorDetails": {
						"type": "string",
						"description": "Optional details for unexpected error responses."
					},
					"errorMessage": {
						"type": "string",
						"description": "Simple message to explain client developers the reason for error."
					},
					"status": {
						"$ref": "#/components/schemas/ApiDefaultResponseStatusEnum",
						"description": "Response status. OK for successful reponses."
					},
					"validationErrorDetails": {
						"$ref": "#/components/schemas/ApiValidationErrorDetails"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ChainProduct_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/ChainProduct"
					},
					"errorDetails": {
						"type": "string",
						"description": "Optional details for unexpected error responses."
					},
					"errorMessage": {
						"type": "string",
						"description": "Simple message to explain client developers the reason for error."
					},
					"status": {
						"$ref": "#/components/schemas/ApiDefaultResponseStatusEnum",
						"description": "Response status. OK for successful reponses."
					},
					"validationErrorDetails": {
						"$ref": "#/components/schemas/ApiValidationErrorDetails"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_PaginatedList_ChainSemiProduct__": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/PaginatedList_ChainSemiProduct_"
					},
					"errorDetails": {
						"type": "string",
						"description": "Optional details for unexpected error responses."
					},
					"errorMessage": {
						"type": "string",
						"description": "Simple message to explain client developers the reason for error."
					},
					"status": {
						"$ref": "#/components/schemas/ApiDefaultResponseStatusEnum",
						"description": "Response status. OK for successful reponses."
					},
					"validationErrorDetails": {
						"$ref": "#/components/schemas/ApiValidationErrorDetails"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "coffee-be",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/chain-api/data/processing-action/{dbId}": {
			"get": {
				"operationId": "GetProcessingAction",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ChainProcessingAction_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Fetches processing action by its _id.",
				"tags": [
					"Processing action"
				],
				"security": [],
				"parameters": [
					{
						"description": "database _id of the product",
						"in": "path",
						"name": "dbId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/chain-api/data/processing-action": {
			"post": {
				"operationId": "PostProcessingAction",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Inserts or updates a processing action. When inserting fields _id, _rev and docType should not be present.",
				"tags": [
					"Processing action"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChainProcessingAction"
							}
						}
					}
				}
			}
		},
		"/chain-api/data/processing-action/list/product/{productId}/organization/{organizationId}": {
			"get": {
				"operationId": "ListProcessingActionsForProductAndOrganization",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedList_ChainProcessingAction__"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Paginated list of processing action for product.",
				"tags": [
					"Processing action"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "organizationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "skuOnly",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/chain-api/data/processing-action/delete": {
			"post": {
				"operationId": "DeleteProcessingAction",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Deletes semi-product.",
				"tags": [
					"Processing action"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChainProcessingAction"
							}
						}
					}
				}
			}
		},
		"/chain-api/data/product/list": {
			"get": {
				"operationId": "ListProducts",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedList_ChainProduct__"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Paginated list of product.",
				"tags": [
					"Product"
				],
				"security": [],
				"parameters": [
					{
						"description": "sort order ASC or DESC",
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {}
					},
					{
						"description": "query limit",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "query offset",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/chain-api/data/product/list/organization/{organizationId}": {
			"get": {
				"operationId": "ListProductsForOrganization",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedList_ChainProduct__"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Paginated list of products for organization.",
				"tags": [
					"Product"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "organizationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/chain-api/data/product/{dbId}": {
			"get": {
				"operationId": "GetProduct",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ChainProduct_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Fetches product by its _id.",
				"tags": [
					"Product"
				],
				"security": [],
				"parameters": [
					{
						"description": "database _id of the product",
						"in": "path",
						"name": "dbId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/chain-api/data/product/external/{linkId}": {
			"get": {
				"operationId": "GetProductByAFId",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ChainProduct_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Returns product according to external database id.",
				"tags": [
					"Product"
				],
				"security": [],
				"parameters": [
					{
						"description": "external database product id",
						"in": "path",
						"name": "linkId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/chain-api/data/product/external/list": {
			"post": {
				"operationId": "ProductsForIds",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Returns a list of products matching to the list of external ids.\nWarning: Id lists and response list do not necessary match in length.\nif invalid external id is provided in the request's list the response list is shorter.",
				"tags": [
					"Product"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "list of external ids",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"type": "number",
									"format": "double"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/chain-api/data/product": {
			"post": {
				"operationId": "PostProduct",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Inserts or updates product. When inserting fields _id, _rev and docType should not be present.",
				"tags": [
					"Product"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChainProduct"
							}
						}
					}
				}
			}
		},
		"/chain-api/data/product/delete": {
			"post": {
				"operationId": "DeleteProduct",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Deletes product.",
				"tags": [
					"Product"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChainProduct"
							}
						}
					}
				}
			}
		},
...
		"/chain-api/data/semi-product/list/product/{productId}": {
			"get": {
				"operationId": "ListSemiProductsForProduct",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedList_ChainSemiProduct__"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Paginated list of semi-products for product.",
				"tags": [
					"Semi product"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "isBuyable",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"description": "(at producer)",
						"in": "query",
						"name": "isSKU",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"description": "(at end customer)",
						"in": "query",
						"name": "isSKUEndCustomer",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}